koordlet: fix util/perf compilation errors on darwin (#1497)

+++ pkg/koordlet/util/perf/perf_unsupported.go
@@ -27,7 +27,7 @@ func NewPerfCollector(cgroupFile *os.File, cpus []int) (*PerfCollector, error) {
 	return &PerfCollector{}, nil
 }
 
-func GetContainerCyclesAndInstructions(collector *PerfCollector) (uint64, uint64, error) {
+func GetContainerCyclesAndInstructions(collector *PerfCollector) (float64, float64, error) {
 	return 0, 0, nil
 }
